requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncio==3.4.3
beautifulsoup4==4.12.2
billiard==4.2.0
cairocffi==1.6.1
CairoSVG==2.7.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.5
lazy_loader==0.3
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
nvidia-tensorrt==99.0.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
prompt-toolkit==3.0.43
psutil==5.9.7
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
sympy==1.12
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.2
unicorn==2.0.1.post1
urllib3==2.1.0
uvicorn==0.25.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
